+++ app/index.html
@@ -21,7 +21,7 @@
             <div main>
                 <!--<span>Root identifier:</span>-->
                 <!--<span class="identifier-name"></span>-->
-                <button class="add-identifier">Add</button>
+                <!--<button class="add-identifier">Add</button>-->
                 <button class="clear-identifier">Clear</button>
             </div>
             <div class="spacer"></div>
